Tag helper buffer for nested elements with multiple children
## Nesting Tag helpers with multiple children

Tag helpers allow you to nest elements.
However, a limitation is that a parent element can only have one
child.

For example, the output from the following code will ignore the first
element.

  tag.div(id: "header") do
    tag.span("hello")
    tag.span("world")
  end

The above will output the following;

<div id="header"><span>world</span></div>

The first `tag.span("hello")` is ignored because the value is not
captured. The helper can only receive the return value of the block
which is the value of the final expression – `tag.span("world")`.

To allow both expressions to be captured, we need to inject a buffer
capture mechanism into the block.
The modification in this commit provides this.

With this modification, you can write the following syntax where `b`
serves as a buffer to collect both elements.

  tag.div(id: "header") do |b|
    b << tag.span("hello")
    b << false && tag.span("This will be ignored")
    b << tag.span("This will be ignored too") if false
    b << tag.span("world")
  end

Refer to TagHelperTest#test_tag_builder_nested_with_multiple_children_including_false
for an example.

## Potential impact on pre-existing code

We patch ActionView::Helpers::TagHelper#tag_string and send in an
empty Array as the buffer into CaptureHelper#capture, which will
be used as a block argument (`b` above).

This previous implementation was sending in `self`
(the TagBuilder object) as the block argument.
However, this was undocumented and was used only in a single test
case – `TagHelperTest#test_tag_builder_nested`.
Furthermore this block argument was unnecessary since you would
normally get the TagBuilder with the `tag` method anyway.

Since preserving this block argument would require that the syntax
contained a mostly unnecessary argument `_tag`,
I removed `self` from the block arguments. This is based on the
assumption that the previous block argument is unlikely to have
been used in the wild.

  tag.div(id: "header") do |_tag, b|
    b << tag.span("hello")
    b << tag.span("world")
  end